Directory Structure for News application
Creates a separate subdirectory "News" for the List/View news
application, as well as a simple news list file (index.php) and a blank
news view file. This design is *not* compatible with Sessions data, and
is based on the structure of our Surveys app.

ToDo: Add CategoryID as a foreign key to Foreign db table, make a
primary key for Foreign, Populate Foreign with RSS data upon page load,
add Sessions, design news_view partial